# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Course Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List all published courses
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseListResponse"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a new course (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseBody"
      responses:
        "201":
          description: Course created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{courseId}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get course details
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDetail"
    put:
      operationId: updateCourse
      tags: [courses]
      summary: Update course (admin only)
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCourseBody"
      responses:
        "200":
          description: Course updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete course (admin only)
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /courses/{courseId}/modules:
    get:
      operationId: listModules
      tags: [modules]
      summary: List course modules
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of modules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Module"
    post:
      operationId: createModule
      tags: [modules]
      summary: Create a module (admin only)
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateModuleBody"
      responses:
        "201":
          description: Module created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
  /courses/{courseId}/modules/{moduleId}:
    put:
      operationId: updateModule
      tags: [modules]
      summary: Update module (admin only)
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: integer
        - in: path
          name: moduleId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateModuleBody"
      responses:
        "200":
          description: Module updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
    delete:
      operationId: deleteModule
      tags: [modules]
      summary: Delete module (admin only)
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: integer
        - in: path
          name: moduleId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Module deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /courses/{courseId}/modules/{moduleId}/lessons:
    get:
      operationId: listLessons
      tags: [lessons]
      summary: List lessons in a module
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: integer
        - in: path
          name: moduleId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of lessons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lesson"
    post:
      operationId: createLesson
      tags: [lessons]
      summary: Create lesson (admin only)
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: integer
        - in: path
          name: moduleId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLessonBody"
      responses:
        "201":
          description: Lesson created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
  /courses/{courseId}/modules/{moduleId}/lessons/{lessonId}:
    get:
      operationId: getLesson
      tags: [lessons]
      summary: Get lesson content (enrolled users only)
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: integer
        - in: path
          name: moduleId
          required: true
          schema:
            type: integer
        - in: path
          name: lessonId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lesson details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
    put:
      operationId: updateLesson
      tags: [lessons]
      summary: Update lesson (admin only)
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: integer
        - in: path
          name: moduleId
          required: true
          schema:
            type: integer
        - in: path
          name: lessonId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLessonBody"
      responses:
        "200":
          description: Lesson updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
    delete:
      operationId: deleteLesson
      tags: [lessons]
      summary: Delete lesson (admin only)
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: integer
        - in: path
          name: moduleId
          required: true
          schema:
            type: integer
        - in: path
          name: lessonId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lesson deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # Auth schemas
...
    Course:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        thumbnailUrl:
          type: string
          nullable: true
        price:
          type: number
        category:
          type: string
        level:
          type: string
          enum: [beginner, intermediate, advanced]
        status:
          type: string
          enum: [draft, published]
        enrollmentCount:
          type: integer
        moduleCount:
          type: integer
        lessonCount:
          type: integer
        durationMinutes:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, description, price, category, level, status, enrollmentCount, moduleCount, lessonCount, durationMinutes, createdAt, updatedAt]
    CourseDetail:
      allOf:
        - $ref: "#/components/schemas/Course"
        - type: object
          properties:
            modules:
              type: array
              items:
                $ref: "#/components/schemas/Module"
            isEnrolled:
              type: boolean
          required: [isEnrolled]
    CourseListResponse:
      type: object
      properties:
        courses:
          type: array
          items:
            $ref: "#/components/schemas/Course"
        total:
          type: integer
      required: [courses, total]
    CreateCourseBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        thumbnailUrl:
          type: string
          nullable: true
        price:
          type: number
        category:
          type: string
        level:
          type: string
          enum: [beginner, intermediate, advanced]
        status:
          type: string
          enum: [draft, published]
      required: [title, description, price, category, level]
    UpdateCourseBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        thumbnailUrl:
          type: string
          nullable: true
        price:
          type: number
        category:
          type: string
        level:
          type: string
          enum: [beginner, intermediate, advanced]
        status:
          type: string
          enum: [draft, published]

    # Module schemas
    Module:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        order:
          type: integer
        lessons:
          type: array
          items:
            $ref: "#/components/schemas/Lesson"
      required: [id, courseId, title, order]
    CreateModuleBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        order:
          type: integer
      required: [title, order]
    UpdateModuleBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        order:
          type: integer

    # Lesson schemas
    Lesson:
      type: object
      properties:
        id:
          type: integer
        moduleId:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        type:
          type: string
          enum: [video, text, pdf, quiz]
        videoUrl:
          type: string
          nullable: true
        content:
          type: string
          nullable: true
        resourceUrl:
          type: string
          nullable: true
        durationMinutes:
          type: integer
          nullable: true
        order:
          type: integer
        isFree:
          type: boolean
        isCompleted:
          type: boolean
      required: [id, moduleId, title, type, order, isFree, isCompleted]
    CreateLessonBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        type:
          type: string
          enum: [video, text, pdf, quiz]
        videoUrl:
          type: string
          nullable: true
        content:
          type: string
          nullable: true
        resourceUrl:
          type: string
          nullable: true
        durationMinutes:
          type: integer
          nullable: true
        order:
          type: integer
        isFree:
          type: boolean
      required: [title, type, order]
    UpdateLessonBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        type:
          type: string
          enum: [video, text, pdf, quiz]
        videoUrl:
          type: string
          nullable: true
        content:
          type: string
          nullable: true
        resourceUrl:
          type: string
          nullable: true
        durationMinutes:
          type: integer
          nullable: true
        order:
          type: integer
        isFree:
          type: boolean

    # Enrollment schemas
...