# lib/api-zod/src/generated/types/updateCourseBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Course Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateCourseBodyLevel } from "./updateCourseBodyLevel";
import type { UpdateCourseBodyStatus } from "./updateCourseBodyStatus";

export interface UpdateCourseBody {
  title?: string;
  description?: string;
  thumbnailUrl?: string | null;
  price?: number;
  category?: string;
  level?: UpdateCourseBodyLevel;
  status?: UpdateCourseBodyStatus;
}